refactor: 파일 구조 변경
- assignment/enum -> assugnment/enums
