+++ src/main/kotlin/com/barostartbe/domain/assignmenttemplate/controller/AssignmentTemplateController.kt
@@ -1,11 +1,10 @@
 package com.barostartbe.domain.assignmenttemplate.controller
 
-import com.barostartbe.domain.assignment.entity.enum.Subject
+import com.barostartbe.domain.assignment.entity.enums.Subject
 import com.barostartbe.domain.assignmenttemplate.dto.request.AssignmentTemplateCreateReq
 import com.barostartbe.domain.assignmenttemplate.dto.request.AssignmentTemplateUpdateReq
 import com.barostartbe.domain.assignmenttemplate.dto.response.AssignmentTemplateDetailRes
 import com.barostartbe.domain.assignmenttemplate.dto.response.AssignmentTemplateListRes
-import com.barostartbe.domain.assignmenttemplate.dto.response.AssignmentTemplateFileRes
 import com.barostartbe.domain.assignmenttemplate.usecase.AssignmentTemplateCreateUseCase
 import com.barostartbe.domain.assignmenttemplate.usecase.AssignmentTemplateDeleteUseCase
 import com.barostartbe.domain.assignmenttemplate.usecase.AssignmentTemplateDetailQueryUseCase
