+++ src/main/kotlin/com/barostartbe/domain/assignmenttemplate/dto/request/AssignmentTemplateUpdateReq.kt
@@ -1,6 +1,6 @@
 package com.barostartbe.domain.assignmenttemplate.dto.request
 
-import com.barostartbe.domain.assignment.entity.enum.Subject
+import com.barostartbe.domain.assignment.entity.enums.Subject
 import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.NotNull
