+++ src/main/kotlin/com/barostartbe/domain/assignmenttemplate/dto/response/AssignmentTemplateByGoalRes.kt
@@ -1,6 +1,6 @@
 package com.barostartbe.domain.assignmenttemplate.dto.response
 
-import com.barostartbe.domain.assignment.entity.enum.Subject
+import com.barostartbe.domain.assignment.entity.enums.Subject
 
 // 과제 템플릿 목표별 조회 응답 DTO (과제 생성시 사용)
 data class AssignmentTemplateByGoalRes(
