+++ src/main/kotlin/com/barostartbe/domain/assignmenttemplate/dto/response/AssignmentTemplateListRes.kt
@@ -1,6 +1,6 @@
 package com.barostartbe.domain.assignmenttemplate.dto.response
 
-import com.barostartbe.domain.assignment.entity.enum.Subject
+import com.barostartbe.domain.assignment.entity.enums.Subject
 import io.swagger.v3.oas.annotations.media.Schema
 
 @Schema(description = "과제 템플릿 목록 조회 응답")
