+++ src/main/kotlin/com/barostartbe/domain/assignmenttemplate/entity/AssignmentTemplate.kt
@@ -1,6 +1,6 @@
 package com.barostartbe.domain.assignmenttemplate.entity
 
-import com.barostartbe.domain.assignment.entity.enum.Subject
+import com.barostartbe.domain.assignment.entity.enums.Subject
 import com.barostartbe.domain.mentor.entity.Mentor
 import com.barostartbe.global.common.entity.BaseEntity
 import jakarta.persistence.*
