+++ src/main/kotlin/com/barostartbe/domain/assignmenttemplate/usecase/AssignmentTemplateByGoalQueryUseCase.kt
@@ -1,7 +1,7 @@
 package com.barostartbe.domain.assignmenttemplate.usecase
 
 
-import com.barostartbe.domain.assignment.entity.enum.Subject
+import com.barostartbe.domain.assignment.entity.enums.Subject
 import com.barostartbe.domain.assignmenttemplate.dto.response.AssignmentTemplateByGoalRes
 import com.barostartbe.domain.assignmenttemplate.dto.response.AssignmentTemplateByGoalRes.TemplateRes
 import com.barostartbe.domain.assignmenttemplate.dto.response.AssignmentTemplateByGoalRes.TemplateFileRes
