+++ src/main/kotlin/com/barostartbe/domain/assignmenttemplate/usecase/AssignmentTemplateListQueryUseCase.kt
@@ -1,6 +1,6 @@
 package com.barostartbe.domain.assignmenttemplate.usecase
 
-import com.barostartbe.domain.assignment.entity.enum.Subject
+import com.barostartbe.domain.assignment.entity.enums.Subject
 import com.barostartbe.domain.assignmenttemplate.dto.response.AssignmentTemplateListRes
 import com.barostartbe.domain.assignmenttemplate.repository.AssignmentTemplateFileRepository
 import com.barostartbe.domain.assignmenttemplate.repository.AssignmentTemplateRepository
