+++ src/test/kotlin/com/barostartbe/domain/assignmenttemplate/usecase/AssignmentTemplateCreateUseCaseTest
@@ -8,7 +8,7 @@ import com.barostartbe.domain.assignmenttemplate.entity.AssignmentTemplate
 import com.barostartbe.domain.assignmenttemplate.entity.AssignmentTemplateFile
 import com.barostartbe.domain.assignmenttemplate.repository.AssignmentTemplateFileRepository
 import com.barostartbe.domain.assignmenttemplate.repository.AssignmentTemplateRepository
-import com.barostartbe.domain.assignment.entity.enum.Subject
+import com.barostartbe.domain.assignment.entity.enums.Subject
 import com.barostartbe.domain.assignmenttemplate.usecase.AssignmentTemplateCreateUseCase
 import com.barostartbe.domain.mentor.entity.Mentor
 import io.kotest.core.spec.style.DescribeSpec
